Restructuring files to use consistent name space
