+++ bambora-client.gemspec
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'bambora/client/version'
+require 'bambora/client'
 
 Gem::Specification.new do |spec|
   spec.name          = 'bambora-client'
