+++ lib/bambora/adapters/json_response.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 module Bambora
-  ##
-  # Parses a JSON response into a Hash
-  class JSONResponse < Response; end
+  module Adapters
+    ##
+    # Parses a JSON response into a Hash
+    class JSONResponse < Response; end
+  end
 end
