+++ lib/bambora/adapters/query_string_response.rb
@@ -1,20 +1,22 @@
 # frozen_string_literal: true
 
 module Bambora
-  ##
-  # Parses a query string response into a Hash
-  class QueryStringResponse < Response
-    def to_h
-      parsed_response = super
-      return error_response if parsed_response.values.flatten.empty? # We didn't get a query string back.
+  module Adapters
+    ##
+    # Parses a query string response into a Hash
+    class QueryStringResponse < Response
+      def to_h
+        parsed_response = super
+        return error_response if parsed_response.values.flatten.empty? # We didn't get a query string back.
 
-      parsed_response.each_with_object({}) { |(key, val), obj| obj[key] = val.length == 1 ? val.first : val }
-    end
+        parsed_response.each_with_object({}) { |(key, val), obj| obj[key] = val.length == 1 ? val.first : val }
+      end
 
-    private
+      private
 
-    def parser
-      CGI
+      def parser
+        CGI
+      end
     end
   end
 end
