+++ lib/bambora/bank/adapters/payment_profile_response.rb
@@ -21,7 +21,7 @@ module Bambora
         private
 
         def parsed_query_string
-          Bambora::QueryStringResponse.new(response).to_h
+          Bambora::Adapters::QueryStringResponse.new(response).to_h
         end
 
         def transform(camel_case_word)
