--- lib/bambora/client/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Bambora
-  class Client
-    VERSION = '0.1.0'
-  end
-end
