+++ spec/bambora/v1/batch_payment_resource_spec.rb
@@ -30,7 +30,7 @@ module Bambora
         }
       end
       let(:client) do
-        Bambora::JSONClient.new(
+        Bambora::Rest::JSONClient.new(
           base_url: base_url,
           api_key: api_key,
           merchant_id: merchant_id,
