+++ spec/bambora/v1/payment_resource_spec.rb
@@ -59,7 +59,7 @@ module Bambora
         }
       end
       let(:client) do
-        Bambora::JSONClient.new(
+        Bambora::Rest::JSONClient.new(
           base_url: base_url,
           merchant_id: merchant_id,
           sub_merchant_id: sub_merchant_id,
